# Services/e_commerce/src/swagger.yaml
openapi: 3.0.0
info:
  title: Product API
  version: 1.0.0
  description: Product management for the E-Commerce platform

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Product:
      type: object
      properties:
        _id:
          type: string
          example: "64b5f8c12d4a2b1234567890"
        title:
          type: string
          example: "MacBook Pro M3"
        price:
          type: number
          example: 2500
        description:
          type: string
          example: "The latest laptop from Apple."
        stock:
          type: number
          example: 50
        images_URL:
          type: array
          items:
            type: string
          example: ["https://res.cloudinary.com/.../img1.jpg", "https://res.cloudinary.com/.../img2.jpg"]
        company:
          type: string
          description: The Company ID (Owner)
          example: "64b5f8c12d4a2b0000000000"
        is_active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time

    APIResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          type: object

paths:
  # --------------------------------------------------------------------------
  # 1. CREATE PRODUCT (Multipart/Form-Data)
  # --------------------------------------------------------------------------
  /api/v1/products/create:
    post:
      summary: Create a new product
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - price
              properties:
                title:
                  type: string
                price:
                  type: number
                description:
                  type: string
                stock:
                  type: number
                category:
                  type: string
                # The File Upload Field
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Select up to 5 image files (jpg, png, webp)
      responses:
        201:
          description: Product Created Successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Product'

  # --------------------------------------------------------------------------
  # 2. GET ALL PRODUCTS (Query Params)
  # --------------------------------------------------------------------------
  /api/v1/products:
    get:
      summary: Get all products (Filter, Sort, Paginate)
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Items per page
        - in: query
          name: sort
          schema:
            type: string
            example: "-price,title"
          description: Sort fields (use minus for descending)
        - in: query
          name: fields
          schema:
            type: string
            example: "title,price"
          description: Select specific fields only
        - in: query
          name: price[gte]
          schema:
            type: number
          description: Filter price >= value
        - in: query
          name: price[lte]
          schema:
            type: number
          description: Filter price <= value
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'

  # --------------------------------------------------------------------------
  # 3. GET PRODUCT BY ID
  # --------------------------------------------------------------------------
  /api/v1/products/{id}:
    get:
      summary: Get a single product by ID
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Product ID
      responses:
        200:
          description: Product details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Product'
        404:
          description: Product not found

  # --------------------------------------------------------------------------
  # 4. UPDATE PRODUCT (Soft Delete included)
  # --------------------------------------------------------------------------
  /api/v1/products/update/{id}:
    patch:
      summary: Update a product (or Soft Delete)
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                price:
                  type: number
                stock:
                  type: number
                is_active:
                  type: boolean
                  description: Set to false to "Soft Delete" the product
      responses:
        200:
          description: Product Updated OR Deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - properties:
                      message:
                        type: string
                        example: "Product updated successfully"
                      data:
                        $ref: '#/components/schemas/Product'